Extract the serialisation test to a separate file
